Update paypal_controller.rb
We do this to avoid a "The totals of the cart item amounts do not match order amounts" error from Paypal.
If not, the tax adjustment of our item is going to be included instead of being skipped (it is an eligible item but not an additional tax).  Our tax rate is already included in the price and it doesn't add to the total amount, and Paypal complains about the mismatch of the sum of the items compared to the total.

At the moment of forking this gem, there were other issues from other people regarding the adjustments. It could be that our tax_adjustment shouldn't be eligible for some reason, but I suspect that the bug comes from some change between the branches 2.2 and 2.3.

I didn't wont to dig deeper into this, since:

- The logic for adjustments is undergoing a huge refactor, which is probably affecting our branch spree/spree#3567
- We also have to introduce a bugfix for the paypal confirmation callback, which was miserably crashing. This means that this gem wasn't properly tested with the 2.3 branch.

At this point on time forking seems a right option.